Added estimate fare and choose location buttons, some enforcement of user input, improvements to XML

+++ app/src/main/java/comcmput301f16t01/github/carrier/FareCalculator.java
@@ -16,7 +16,7 @@ public class FareCalculator {
         // TODO calculate a fare. (this is not for project part 2)
     }
 
-    public int getEstimate() {
+    public double getEstimate() {
         return 0;
     }
 }
